bdev: Fix QoS channel destroy
QoS holds a ref to the channel in qos->ch
qos_cache belongs to the channel, so last cache
instance is destroyed when qos->ch is put (assuming
there are no more refs). This only happens when
QoS is fully destroyed (bdev_close). That meas that
`assert(TAILQ_EMPTY(&qos->cache_list));` was not correct,
it is true only when qos->ch is NULL
Next, if we put qos->ch and free(qos) then we'll have
the heap-use-after-free in bdev_qos_cache_destroy. To
fix it, don't free qos in bdev_qos_channel_destroy but
set qos->ch = NULL indicating that bdev_qos_cache_destroy
must free its memory

Signed-off-by: Alexey Marchuk <alexeymar@nvidia.com>
Change-Id: I87d0874a211a51273055bca9e7b9856aaebd5a55