sock: stop returning 1 on spdk_sock_posix_fd_connect*
It was a reasonable hint for the user to try another address if
available; however, it complicates usage and error handling. It is
preferable to continue on failure if another address is available.
In the case of a critical failure, it will be encountered again quickly.

Additionally, by returning 1, we lose the actual error code. This change
is also a prework for a future change where, instead of returning -1, a
more specific errno will be returned.

Change-Id: Ia6fa3c6af5f80129adb708dbd5971a3711a9b6bd
Signed-off-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26106
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Tomasz Zawadzki <tomasz@tzawadzki.com>