sgl: add spdk_iov_sgl_iovcnt()
In most, if not all, cases where spdk_iov_sgl is used, we need to take
the number of iovs filled with data, so have to do something like:

  iovcnt = SPDK_COUNTOF(some_iovs) - sgl.iovcnt;

which isn't really obvious and can lead to bugs.  Instead, let's
remember the size of the iov array and have a helper that returns
iovcnt.

Change-Id: I6d14a6840c4aca424272f85ab15cfd6a4581ccf2
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>