pkgdep/git: Add support for installing DPDK's contigmem driver
The main goal here is to arm freebsd environment with ability to
load aforementioned driver early on during kernel's bootup. This
should mitigate potential issues with allocating enough
contiguous memory space for DPDK during tests.

dpdk-stable's latest branch is used by default as a source for
the driver.

As we are at it, small reshuffling of the freebsd_*() functions
is done to make sure it's clear what they are responsible for.

CI autorun workflows should utilize this driver under FreeBSD in
the following manner:

  # During qcow2 build
  $ sudo  ./autotest_setup.sh  [...] --test-conf="contigmem"

The above will build, install and prep the driver for the loader -
default of 2GB will be requested on boot which should be enough
and match default requirement by autotest.

  # during tests under FreeBSD
  export SKIP_HUGE=yes USE_INSTALLED_CONTIGMEM=yes
  ...

The SKIP_HUGE flag is meant for setup.sh to avoid any tweaking
of contigmem buffers; USE_INSTALLED_CONTIGMEM makes sure that
pre-installed driver won't be overwritten by autobuild's.

Change-Id: I16216d9b17144d046cbcffa9d66783d9b8e91102
Signed-off-by: Michal Berger <michal.berger@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26376
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Tomasz Zawadzki <tomasz@tzawadzki.com>
Reviewed-by: Konrad Sztyber <ksztyber@nvidia.com>
Community-CI: Mellanox Build Bot