lib/util/lut: introduced spdk_lut_reset
The spdk_lut_reset API resets the LUT and resizes it to the original
size.

This API is useful, for instance, when a LUT is being reused after it
has been expanded and is utilizing a lot of memory for LUT nodes, with
all the nodes residing in its free_nodes list.

Change-Id: Ic28b7819c6fa26a48e0a6368e93b5bd59799ca59
Signed-off-by: Anton Nayshtut <anayshtut@nvidia.com>